requirements.txt
RPi.GPIO==0.6.3
RPIO==0.10.0
